+++ components/PopupReview/index.js
@@ -1,131 +1,21 @@
-import { useEffect, useState } from "react";
-import { FaStar } from "react-icons/fa";
-import Image from "next/image";
+import PopupInputReview from "../PopupInputReview";
+import PopupReviewSucces from "../PopupReviewSucces";
+import { useSelector } from "react-redux";
 
-function PopupReview() {
-  const [currentValue, setCurrentValue] = useState(5);
-  const [hoverValue, setHoverValue] = useState(undefined);
-  const stars = Array(5).fill(0);
-
-  const [userClicked, setUserClicked] = useState(false);
-
-  const handleClick = (value) => {
-    setCurrentValue(value);
-    setUserClicked(true);
-  };
-
-  const handleMouseOver = (newHoverValue) => {
-    setHoverValue(newHoverValue);
-  };
-
-  const handleMouseLeave = () => {
-    setHoverValue(undefined);
-  };
-  useEffect(() => {
-    console.log(userClicked);
-  }, [userClicked]);
+const PopupReview = () => {
+  const showPopupReview = useSelector((state) => state.reviewQuery.togglePopupReview);
 
   return (
-    <div style={styles.container}>
-      <img
-        src="/close-chat.svg"
-        alt="close"
-        className="absolute"
-        style={{ top: "22px", right: "22px" }}
-      />
-      <h2 className="text-2xl font-semibold" style={{ marginBottom: "22px" }}>
-        Tulis Review
-      </h2>
-      <Image
-        src="/building.svg"
-        alt="profile"
-        width={81}
-        height={81}
-        className="rounded-full object-cover"
-      />
-      <p className="mt-2.5 text-base" style={{ marginBottom: "22px" }}>
-        name
-      </p>
-      <div style={styles.stars}>
-        {stars.map((_, index) => {
-          return (
-            <FaStar
-              key={index}
-              size={24}
-              onClick={() => handleClick(index + 1)}
-              onMouseOver={() => handleMouseOver(index + 1)}
-              onMouseLeave={handleMouseLeave}
-              color={(hoverValue || currentValue) > index ? colors.orange : colors.grey}
-              style={{
-                marginRight: 10,
-                cursor: "pointer",
-              }}
-            />
-          );
-        })}
-      </div>
-      <textarea
-        placeholder="Additional Comments"
-        onChange={() => {
-          setUserClicked(true);
-        }}
-        className="bg-grey text-secondary focus:outline-blue"
-        style={styles.textarea}
-      />
-
-      {userClicked ? (
-        <button style={styles.button} className="bg-blue">
-          Submit Review
-        </button>
-      ) : (
-        <button style={styles.button} className= "bg-secondary" disabled>
-          Submit Review
-        </button>
-      )}
-    </div>
+    <>
+      {
+        showPopupReview &&
+        <div className="absolute top-0 left-0 bg-secondary w-full h-full flex justify-center items-center">
+          <PopupInputReview />
+          <PopupReviewSucces />
+        </div>
+      }
+    </>
   );
-}
-
-const styles = {
-  container: {
-    display: "flex",
-    flexDirection: "column",
-    alignItems: "center",
-    justifyContent: "center",
-    width: "339px",
-    height: "456px",
-    position: "relative",
-  },
-  stars: {
-    display: "flex",
-    flexDirection: "row",
-  },
-  textarea: {
-    border: "1px solid #a9a9a9",
-    borderRadius: 5,
-    padding: 10,
-    margin: "20px 0",
-    minHeight: 54,
-    width: 236,
-    border: "none",
-    fontSize: "8px",
-  },
-  button: {
-    border: "none",
-    borderRadius: 6,
-    color: "#fff",
-    width: 235,
-    height: 34,
-    textAlign: "center",
-    fontSize: 12,
-  },
-  buttonBlue: {
-    backgroundColor: "#197BEB",
-  },
-};
-const colors = {
-  orange: "#FFBA5A",
-  grey: "#a9a9a9",
 };
 
 export default PopupReview;
