+++ redux/store.js
@@ -1,8 +1,10 @@
 import { configureStore } from "@reduxjs/toolkit";
 import  MediaQueryReducer  from "./features/MediaQuerySlice";
+import  ReviewQueryReducer  from "./features/ReviewQuerySlice";
 
 export default configureStore({
   reducer: {
-    mediaQuery: MediaQueryReducer
+    mediaQuery: MediaQueryReducer,
+    reviewQuery: ReviewQueryReducer
   }
 })
